CalcData modifications
Rewrote calcData from scratch for two reasons. 1) to make the output more useful. Previously the output was more specific. It would construct a DF with columns that give the horizontal analysis across years. This had drawbacks because of the pervasiveness of "n/a" values. Now the DF columns are just the flat value of that year and not the change in value. This makes "n/a" values less impactful because we can calculate long term changes in value with only two datapoints. 2) to execute the task in a pandas-oriented way. Now there are much fewer if statements and arrays because the code uses pandas methods to accomplish its task.