Class abstraction and renaming
Provided a class interface for working with horizontal profiles. Renamed clacData to be more specific. Added horizontalCalc to move the analysis forward from values -> change in values. DFs consisting of changes in values will be represented by a new class. Would be smart to write a super class that has the base methods for both horizontalprofile and horizontalAnalysis classes.